+++ packages/hydrated_bloc/test/hydrated_storage_test.dart
@@ -2,181 +2,134 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:hive/hive.dart';
 import 'package:hydrated_bloc/hydrated_bloc.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as p;
-import 'package:path_provider/path_provider.dart';
-import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
 import 'package:pedantic/pedantic.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
+// ignore: implementation_imports
+import 'package:hive/src/hive_impl.dart';
 
 class MockBox extends Mock implements Box<dynamic> {}
 
-class MockPathProviderPlatform extends Mock
-    with MockPlatformInterfaceMixin
-    implements PathProviderPlatform {
-  MockPathProviderPlatform({
-    @required this.temporaryPath,
-    @required this.getTemporaryPathCalled,
-  });
-  final String temporaryPath;
-  final VoidCallback getTemporaryPathCalled;
-
-  @override
-  Future<String> getTemporaryPath() async {
-    getTemporaryPathCalled();
-    return temporaryPath;
-  }
-}
-
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-
   group('HydratedStorage', () {
     final cwd = Directory.current.absolute.path;
-    var getTemporaryDirectoryCallCount = 0;
+    final storageDirectory = Directory(cwd);
 
-    setUp(() async {
-      PathProviderPlatform.instance = MockPathProviderPlatform(
-        temporaryPath: cwd,
-        getTemporaryPathCalled: () => ++getTemporaryDirectoryCallCount,
-      );
-    });
-
-    Storage storage;
+    late Storage storage;
 
     tearDown(() async {
-      await storage?.clear();
+      await storage.clear();
     });
 
     group('migration', () {
       test('returns correct value when file exists', () async {
-        final directory = await getTemporaryDirectory();
-        File('${directory.path}/.hydrated_bloc.json')
+        File('${storageDirectory.path}/.hydrated_bloc.json')
           ..writeAsStringSync(json.encode({
             'CounterBloc': json.encode({'value': 4})
           }));
-        storage = await HydratedStorage.build();
+        storage = await HydratedStorage.build(
+          storageDirectory: storageDirectory,
+        );
         expect(storage.read('CounterBloc')['value'] as int, 4);
       });
     });
 
     group('build', () {
       setUp(() async {
-        await (await HydratedStorage.build()).clear();
-        getTemporaryDirectoryCallCount = 0;
-      });
-
-      test('calls getTemporaryDirectory when storageDirectory is null',
-          () async {
-        storage = await HydratedStorage.build();
-        expect(getTemporaryDirectoryCallCount, 1);
-      });
-
-      test(
-          'does not call getTemporaryDirectory '
-          'when storageDirectory is null and kIsWeb', () async {
-        HydratedStorage.isWeb = true;
-        final completer = Completer<void>();
-        await runZoned(() {
-          HydratedStorage.build().whenComplete(completer.complete);
-          return completer.future;
-        }, onError: (Object _) {});
-        HydratedStorage.isWeb = kIsWeb;
-        expect(getTemporaryDirectoryCallCount, 0);
-      });
-
-      test(
-          'does not call getTemporaryDirectory '
-          'when storageDirectory is defined', () async {
-        storage = await HydratedStorage.build(storageDirectory: Directory(cwd));
-        expect(getTemporaryDirectoryCallCount, 0);
+        await (await HydratedStorage.build(storageDirectory: storageDirectory))
+            .clear();
       });
 
       test('reuses existing instance when called multiple times', () async {
-        final instanceA = storage = await HydratedStorage.build();
-        final beforeCount = getTemporaryDirectoryCallCount;
-        final instanceB = await HydratedStorage.build();
-        final afterCount = getTemporaryDirectoryCallCount;
-        expect(beforeCount, afterCount);
+        final instanceA = storage = await HydratedStorage.build(
+          storageDirectory: storageDirectory,
+        );
+        final instanceB = await HydratedStorage.build(
+          storageDirectory: storageDirectory,
+        );
         expect(instanceA, instanceB);
       });
 
       test('creates internal HiveImpl with correct directory', () async {
-        storage = await HydratedStorage.build();
-        final box = HydratedStorage.hive?.box<dynamic>('hydrated_box');
-        final directory = await getTemporaryDirectory();
+        storage = await HydratedStorage.build(
+          storageDirectory: storageDirectory,
+        );
+        final box = HydratedStorage.hive.box<dynamic>('hydrated_box');
         expect(box, isNotNull);
-        expect(box.path, p.join(directory.path, 'hydrated_box.hive'));
+        expect(box.path, p.join(storageDirectory.path, 'hydrated_box.hive'));
       });
     });
 
     group('default constructor', () {
       const key = '__key__';
       const value = '__value__';
-      Box box;
+      late Box box;
 
       setUp(() {
         box = MockBox();
+        when(box).calls(#deleteFromDisk).thenReturn(Future<void>.value());
         storage = HydratedStorage(box);
       });
 
       group('read', () {
         test('returns null when box is not open', () {
-          when(box.isOpen).thenReturn(false);
+          when(box).calls(#isOpen).thenReturn(false);
           expect(storage.read(key), isNull);
         });
 
         test('returns correct value when box is open', () {
-          when(box.isOpen).thenReturn(true);
-          when<dynamic>(box.get(any)).thenReturn(value);
+          when(box).calls(#isOpen).thenReturn(true);
+          when(box).calls(#get).thenReturn(value);
           expect(storage.read(key), value);
-          verify<dynamic>(box.get(key)).called(1);
+          verify(box).called(#get).withArgs(positional: [key]).once();
         });
       });
 
       group('write', () {
         test('does nothing when box is not open', () async {
-          when(box.isOpen).thenReturn(false);
+          when(box).calls(#isOpen).thenReturn(false);
           await storage.write(key, value);
-          verifyNever(box.put(any, any));
+          verify(box).called(#put).never();
         });
 
         test('puts key/value in box when box is open', () async {
-          when(box.isOpen).thenReturn(true);
+          when(box).calls(#isOpen).thenReturn(true);
+          when(box).calls(#put).thenReturn(Future<void>.value());
           await storage.write(key, value);
-          verify(box.put(key, value)).called(1);
+          verify(box).called(#put).withArgs(positional: [key, value]).once();
         });
       });
 
       group('delete', () {
         test('does nothing when box is not open', () async {
-          when(box.isOpen).thenReturn(false);
+          when(box).calls(#isOpen).thenReturn(false);
           await storage.delete(key);
-          verifyNever(box.delete(any));
+          verify(box).called(#delete).never();
         });
 
         test('puts key/value in box when box is open', () async {
-          when(box.isOpen).thenReturn(true);
+          when(box).calls(#isOpen).thenReturn(true);
+          when(box).calls(#delete).thenReturn(Future<void>.value());
           await storage.delete(key);
-          verify(box.delete(key)).called(1);
+          verify(box).called(#delete).withArgs(positional: [key]).once();
         });
       });
 
       group('clear', () {
         test('does nothing when box is not open', () async {
-          when(box.isOpen).thenReturn(false);
+          when(box).calls(#isOpen).thenReturn(false);
           await storage.clear();
-          verifyNever(box.deleteFromDisk());
+          verify(box).called(#deleteFromDisk).never();
         });
 
         test('deletes box when box is open', () async {
-          when(box.isOpen).thenReturn(true);
+          when(box).calls(#isOpen).thenReturn(true);
           await storage.clear();
-          verify(box.deleteFromDisk()).called(1);
+          verify(box).called(#deleteFromDisk).once();
         });
       });
     });
@@ -213,6 +166,8 @@ void main() {
       final docs = p.join(cwd, 'docs');
 
       tearDown(() async {
+        await storage.clear();
+        await Hive.close();
         await Directory(temp).delete(recursive: true);
         await Directory(docs).delete(recursive: true);
       });
@@ -220,7 +175,7 @@ void main() {
       test('Hive and Hydrated default directories', () async {
         Hive.init(docs);
         storage = await HydratedStorage.build(
-          storageDirectory: Directory(temp),
+          storageDirectory: Directory(temp)..createSync(),
         );
 
         var box = await Hive.openBox<String>('hive');
@@ -228,15 +183,13 @@ void main() {
         expect(box.get('name'), 'hive');
         await Hive.close();
 
+        // https://github.com/hivedb/hive/pull/521#issuecomment-767903897
+        (Hive as HiveImpl).homePath = null;
+
         Hive.init(docs);
         box = await Hive.openBox<String>('hive');
-        try {
-          expect(box.get('name'), isNotNull);
-          expect(box.get('name'), 'hive');
-        } finally {
-          await storage.clear();
-          await Hive.close();
-        }
+        expect(box.get('name'), isNotNull);
+        expect(box.get('name'), 'hive');
       });
     });
   });