+++ packages/hydrated_bloc/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 7.0.0-nullsafety.0
+
+- **BREAKING**: opt into null safety
+- **BREAKING**: refactor: upgrade to `bloc ^7.0.0-nullsafety.1`
+- **BREAKING**: refactor: remove `flutter` dependency
+- **BREAKING**: `storageDirectory` is required when calling `HydratedStorage.build`
+- **BREAKING**: `HydratedCubit.storage` is removed in favor of `HydratedBloc.storage`
+- feat!: upgrade Dart SDK constraints to `>=2.12.0-0 <3.0.0`
+
 # 6.1.0
 
 - feat: export `package:bloc/bloc.dart`